requirements.txt
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
aiohttp==3.9.1
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0